move/refactor ListSelectionHelper
